+++ packages/video_player_avplay/tizen/src/video_player.cc
@@ -199,6 +199,14 @@ void VideoPlayer::SendRestored() {
   }
 }
 
+void VideoPlayer::SendADFromDash(flutter::EncodableMap ad_info) {
+  flutter::EncodableMap result = {
+      {flutter::EncodableValue("event"), flutter::EncodableValue("adFromDash")},
+      {flutter::EncodableValue("adInfo"), flutter::EncodableValue(ad_info)},
+  };
+  PushEvent(flutter::EncodableValue(result));
+}
+
 void VideoPlayer::SendError(const std::string &error_code,
                             const std::string &error_message) {
   if (event_sink_) {