+++ packages/video_player_avplay/tizen/src/video_player.h
@@ -68,6 +68,7 @@ class VideoPlayer {
   virtual flutter::EncodableMap GetData(const flutter::EncodableList &data) {
     return flutter::EncodableMap{};
   }
+  virtual bool UpdateToken(const std::string &token) { return false; };
   virtual flutter::EncodableList GetActiveTrackInfo() {
     return flutter::EncodableList{};
   }
@@ -87,6 +88,7 @@ class VideoPlayer {
   void SendPlayCompleted();
   void SendIsPlayingState(bool is_playing);
   void SendRestored();
+  void SendADFromDash(flutter::EncodableMap ad_info);
   void SendError(const std::string &error_code,
                  const std::string &error_message);